# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Awesome API",
        "description": "Swagger",
        "contact": {
            "email": "your@email.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all users for admin",
                "operationId": "9ae473e4762fd14440b777120aa3f205",
                "responses": {
                    "200": {
                        "description": "A list of all users"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Add a new user",
                "operationId": "6553524bf96358d6d0af5ecba02714d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "first_name": {
                                            "type": "string"
                                        },
                                        "last_name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "password": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/approve/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Approve a user request",
                "operationId": "bbaa9432bad0246564492e3ee41961bd",
                "responses": {
                    "200": {
                        "description": "User approved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/reject/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Reject a user request",
                "operationId": "a4dd32c400c32f5fc7ec0160a3ad5bf3",
                "responses": {
                    "200": {
                        "description": "User reject successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/search/{keyword}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Search users by keyword",
                "operationId": "92885853bcf78aaade3681d5de2e92ad",
                "responses": {
                    "200": {
                        "description": "A list of all users"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete a user by ID",
                "operationId": "91c9bfbf1bbf4d24b50a5b3fe5945b37",
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "confirm_password",
                                    "role_id"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Mohammad"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Salem"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "mohammad@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    },
                                    "confirm_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    },
                                    "role_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "mohammad@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login successfully"
                    },
                    "403": {
                        "description": "Role mismatch"
                    },
                    "401": {
                        "description": "Incorrect email or password"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout the authenticated user",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/forgot-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Send password reset link to email",
                "operationId": "b93bd61884f12f561e37f8459dd3a625",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "albashaosayd@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reset link sent to your email."
                    },
                    "500": {
                        "description": "Unable to send reset link."
                    }
                }
            }
        },
        "/api/verify-2fa": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verify 2FA code",
                "operationId": "ef04a8108d42942a7bb9fb8f73a9e2ff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "code"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "2FA Verified Successfully"
                    }
                }
            }
        },
        "/api/resetPassword": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Reset user password using token",
                "operationId": "66d0a5c892b4deff44faa43cf7c4e86a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "abcdef123456"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "osayd@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password has been reset"
                    },
                    "500": {
                        "description": "Reset failed"
                    }
                }
            }
        },
        "/api/properties": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get all properties",
                "operationId": "93f6f24283a4d0eefb208f5481a56a73",
                "responses": {
                    "200": {
                        "description": "A list of properties"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Admin",
            "description": "Admin"
        },
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Properties",
            "description": "Properties"
        }
    ]
}